Move src files out of unnecessary promise directory and rename Promise.js to CancellablePromise.js
