+++ src/CancellablePromise.js
